backend/requirements.txt
python-dotenv
fastapi==0.119.0
fastapi-cli==0.0.13
pymongo==4.15.4
python-dotenv==1.2.1
pydantic==2.12.4